ui/cosmic_styles.py
#!/usr/bin/env python3
"""
F.A.M.E. 11.0 - Premium Cosmic Styling
2028 futuristic design system
"""

from typing import Dict, Any, List

# Try customtkinter
try:
    import customtkinter as ctk
    CTK_AVAILABLE = True
except ImportError:
    CTK_AVAILABLE = False
    import tkinter as tk
    from tkinter import ttk


class CosmicTheme:
    """Premium 2028 cosmic theme system"""
    
    @staticmethod
    def setup_cosmic_theme():
        """Setup cosmic theme"""
        cosmic_colors = {
            "cosmic_dark": {
                "bg": "#0a0a1a",
                "fg": "#00ffff", 
                "accent": "#ff00ff",
                "secondary": "#ffff00",
                "success": "#00ff00",
                "warning": "#ffaa00",
                "error": "#ff4444"
            },
            "quantum_blue": {
                "bg": "#001122",
                "fg": "#00ffff",
                "accent": "#0088ff", 
                "secondary": "#00ff88",
                "success": "#00ff00",
                "warning": "#ffcc00",
                "error": "#ff0066"
            },
            "neon_matrix": {
                "bg": "#000000",
                "fg": "#00ff00", 
                "accent": "#ff00ff",
                "secondary": "#00ffff",
                "success": "#00ff00",
                "warning": "#ffff00",
                "error": "#ff0000"
            }
        }
        return cosmic_colors
    
    @staticmethod
    def create_cosmic_button(parent, text, command, **kwargs):
        """Create a cosmic-style button"""
        if CTK_AVAILABLE:
            return ctk.CTkButton(
                parent,
                text=text,
                command=command,
                font=ctk.CTkFont(family="Segoe UI", size=14, weight="bold"),
                border_width=2,
                border_color="#00ffff",
                fg_color="transparent",
                hover_color="#00ffff",
                text_color="#00ffff",
                **kwargs
            )
        else:
            return tk.Button(
                parent,
                text=text,
                command=command,
                bg="#0a0a1a",
                fg="#00ffff",
                font=("Arial", 12, "bold"),
                **kwargs
            )


class AnimatedBackground:
    """Animated cosmic background"""
    
    def __init__(self, parent):
        self.parent = parent
        if CTK_AVAILABLE:
            self.canvas = tk.Canvas(parent, highlightthickness=0, bg="#0a0a1a")
        else:
            self.canvas = tk.Canvas(parent, highlightthickness=0, bg="#0a0a1a")
        self.canvas.pack(fill="both", expand=True)
        self.stars = []
    
    def start_animations(self):
        """Start background animations"""
        self._create_stars()
        self._animate_particles()
    
    def _create_stars(self):
        """Create starfield background"""
        import random
        width = self.canvas.winfo_width() or 1600
        height = self.canvas.winfo_height() or 900
        
        for _ in range(100):
            x = random.randint(0, width)
            y = random.randint(0, height)
            size = random.uniform(0.5, 2)
            star = self.canvas.create_oval(x, y, x+size, y+size, fill="white", outline="")
            self.stars.append(star)
    
    def _animate_particles(self):
        """Animate cosmic particles"""
        pass

