requirements_production.txt
# FAME Production Requirements
# Enterprise-grade dependencies for production-ready AI assistant

# Core Python

# Async & Networking
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1
yfinance>=0.2.28
ccxt>=4.0.0
pyyaml>=6.0.0
pyarrow>=12.0.0
tqdm>=4.65.0

# System Monitoring
psutil>=5.9.0

# GUI (Desktop) # Usually comes with Python
Pillow>=10.0.0

# Voice Processing
speechrecognition>=3.10.0
pyttsx3>=2.90
# pyaudio - Install separately based on OS

# Docker Integration
docker>=6.1.0
python-dotenv>=1.0.0

# Financial Data & Analysis
yfinance>=0.2.28
vaderSentiment>=3.3.2

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Machine Learning (Optional - for advanced features)
# transformers>=4.30.0
scikit-learn>=1.3.0

# LLM Integration (Optional - for dynamic reasoning)
openai>=1.0.0  # For dynamic reasoning and understanding

chromadb>=0.4.0  # Vector database for memory
# sentence-transformers>=2.2.0  # Embeddings for semantic search (install on-demand to avoid heavy torch wheels)

# Logging & Monitoring
structlog>=23.1.0  # Structured logging
prometheus-client>=0.17.0  # Metrics exporter

# API & Serving
fastapi>=0.110.0
uvicorn[standard]>=0.23.0

# Distributed Tracing (Optional)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0

# Security
cryptography>=41.0.0  # For secure data handling

# Configuration Management
pyyaml>=6.0

